Move lib-testing code to works.bosk.libtesting
