+++ bosk-core/src/main/java/module-info.java
@@ -24,6 +24,6 @@ module works.bosk.core {
 	// fully decided what parts of these packages should be a permanent part of
 	// the core bosk API.
 	exports works.bosk.bytecode to works.bosk.jackson;
-	exports works.bosk.util to works.bosk.jackson, works.bosk.mongo, works.bosk.testing;
+	exports works.bosk.util to works.bosk.jackson, works.bosk.mongo, works.bosk.testing, works.bosk.libtesting;
 	exports works.bosk.logging to works.bosk.logback, works.bosk.mongo, works.bosk.sql, works.bosk.testing; // May be ok to export this, but give it a think first
 }
