+++ bosk-core/src/test/java/works/bosk/MapValueTest.java
@@ -16,6 +16,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import works.bosk.libtesting.AbstractBoskTest;
 
 import static java.util.Arrays.asList;
 import static java.util.Collections.emptyMap;
