+++ bosk-core/src/test/java/works/bosk/ReferenceTest.java
@@ -8,6 +8,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import works.bosk.annotations.ReferencePath;
 import works.bosk.exceptions.InvalidTypeException;
+import works.bosk.libtesting.AbstractBoskTest;
 
 import static java.util.Collections.emptyList;
 import static java.util.Collections.singletonList;
