+++ bosk-core/src/test/java/works/bosk/dereferencers/PathCompilerTest.java
@@ -14,7 +14,6 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DynamicTest;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestFactory;
-import works.bosk.AbstractBoskTest;
 import works.bosk.Bosk;
 import works.bosk.Bosk.NonexistentEntryException;
 import works.bosk.BoskConfig;
@@ -28,9 +27,10 @@ import works.bosk.Path;
 import works.bosk.Reference;
 import works.bosk.SideTable;
 import works.bosk.StateTreeNode;
-import works.bosk.TestEntityBuilder;
 import works.bosk.exceptions.InvalidTypeException;
 import works.bosk.exceptions.NonexistentReferenceException;
+import works.bosk.libtesting.AbstractBoskTest;
+import works.bosk.libtesting.TestEntityBuilder;
 
 import static java.util.Collections.singletonList;
 import static org.junit.jupiter.api.Assertions.assertEquals;
