+++ bosk-core/src/test/java/works/bosk/drivers/BsonRoundTripConformanceTest.java
@@ -3,7 +3,7 @@ package works.bosk.drivers;
 import org.junit.jupiter.api.BeforeEach;
 import works.bosk.testing.drivers.DriverConformanceTest;
 
-import static works.bosk.AbstractRoundTripTest.bsonRoundTripFactory;
+import static works.bosk.libtesting.AbstractRoundTripTest.bsonRoundTripFactory;
 
 /**
  * TODO: Move to bosk-mongo
