+++ bosk-jackson/src/test/java/works/bosk/jackson/JacksonSerializerTest.java
@@ -22,7 +22,6 @@ import tools.jackson.databind.JavaType;
 import tools.jackson.databind.ObjectMapper;
 import tools.jackson.databind.json.JsonMapper;
 import tools.jackson.databind.type.TypeFactory;
-import works.bosk.AbstractBoskTest;
 import works.bosk.BindingEnvironment;
 import works.bosk.Bosk;
 import works.bosk.Catalog;
@@ -37,13 +36,14 @@ import works.bosk.Reference;
 import works.bosk.SideTable;
 import works.bosk.StateTreeNode;
 import works.bosk.TaggedUnion;
-import works.bosk.TestEntityBuilder;
 import works.bosk.annotations.DeserializationPath;
 import works.bosk.annotations.Polyfill;
 import works.bosk.annotations.ReferencePath;
 import works.bosk.exceptions.MalformedPathException;
 import works.bosk.exceptions.ParameterUnboundException;
 import works.bosk.exceptions.UnexpectedPathException;
+import works.bosk.libtesting.AbstractBoskTest;
+import works.bosk.libtesting.TestEntityBuilder;
 
 import static java.util.Arrays.asList;
 import static java.util.Collections.emptyMap;
