+++ lib-testing/src/main/java/module-info.java
@@ -1,0 +1,18 @@
+/**
+ * Internal test infrastructure for Bosk subprojects.
+ */
+module works.bosk.libtesting {
+	requires works.bosk.core;
+	requires works.bosk.testing;
+	requires works.bosk.jackson;
+	requires works.bosk.mongo;
+	requires works.bosk.bosonSerializer;
+
+	requires org.slf4j;
+	requires org.junit.jupiter.api;
+	requires org.junit.jupiter.params;
+
+	requires static lombok;
+
+	exports works.bosk.libtesting;
+}
