+++ lib-testing/src/main/java/works/bosk/libtesting/SerializerRoundTripTest.java
@@ -1,7 +1,10 @@
-package works.bosk;
+package works.bosk.libtesting;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
+import works.bosk.Bosk;
+import works.bosk.DriverFactory;
+import works.bosk.Reference;
 import works.bosk.exceptions.InvalidTypeException;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
